Fix collabs tests, add shared functions, move editRepo helpers
