+++ cypress/e2e/dolthub/privatePaths/render/databaseSettings/branchProtection.spec.ts
