+++ cypress/e2e/dolthub/privatePaths/render/databaseSettings/index.spec.ts
